Implemented stack-based navigation
Navigation now works using Navigator.push(), which passes an Item to a new CompassScreen widget. Return to the LocationList screen by either swiping right or pressing the back button on the CompassScreen's AppBar. Since the NavigationBar is not compatible with this Navigator system, it has been removed from both screens.